requirements.txt
azure-functions
sqlalchemy
snowflake-sqlalchemy
snowflake-connector-python
sparkpost
tabulate
pandas